teamtraces & huidige locatie rework, nu 2 in 1 functie. + startgame bug fixing

+++ src/CityCheckAPI/DataLayer/DataLayer/Models/Game.cs
@@ -8,7 +8,7 @@ namespace DataLayer
         public int GameCode { get; set; }
         public int TijdsDuur { get; set; }
         public bool hasStarted { get; set; }
-        public int millisStarted { get; set; }
+        public double millisStarted { get; set; }
         public List<Team> Teams { get; set; }
 
     }
